pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "^2.1.1"}
loguru = "^0.7.0"
sqlalchemy = "^2.0.19"
dependency-injector = {extras = ["yaml"], version = "^4.41.0"}
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
asyncpg = "^0.28.0"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
emails = "^0.6"
web3 = "^6.8.0"
eth-account = "^0.9.0"
httpx = "^0.24.1"
typing-extensions = "^4.7.1"
celery = "^5.3.1"
django-celery-beat = "^2.5.0"
propan = {extras = ["async-rabbit"], version = "^0.1.5.18"}
python-socketio = {extras = ["asyncio-client"], version = "^5.8.0"}
aioredis = "^2.0.1"
redis = "^5.0.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    ".", "src"
]

asyncio_mode="auto"